+++ es6/modules/common.js
@@ -1,5 +1,5 @@
 const wrapper = require("../module-wrapper");
-const ctXML = "[Content_Types].xml";
+const { concatArrays } = require("../doc-utils");
 const docxContentType =
 	"application/vnd.openxmlformats-officedocument.wordprocessingml.document.main+xml";
 const docxmContentType =
@@ -11,54 +11,37 @@ class Common {
 	constructor() {
 		this.name = "Common";
 	}
-	getFileType({ zip, contentTypes, overrides, defaults, doc }) {
-		if (!contentTypes) {
+	set({ invertedContentTypes }) {
+		if (invertedContentTypes) {
+			this.invertedContentTypes = invertedContentTypes;
+		}
+	}
+	getFileType({ doc }) {
+		if (!this.invertedContentTypes) {
 			return;
 		}
 		let fileType = null;
-		const partNames = [];
-		for (let i = 0, len = overrides.length; i < len; i++) {
-			const override = overrides[i];
-			const contentType = override.getAttribute("ContentType");
-			const partName = override.getAttribute("PartName").substr(1);
-			partNames.push(partName);
-			if (contentType === docxContentType) {
-				fileType = "docx";
-				doc.targets.push(partName);
-			}
-			if (contentType === docxmContentType) {
-				fileType = "docx";
-				doc.targets.push(partName);
-			}
-			if (contentType === pptxContentType) {
-				fileType = "pptx";
-				doc.targets.push(partName);
-			}
+		const invertedContentTypes = this.invertedContentTypes;
+		if (invertedContentTypes[docxContentType]) {
+			fileType = "docx";
+			doc.targets = concatArrays([
+				doc.targets,
+				invertedContentTypes[docxContentType],
+			]);
 		}
-		if (fileType) {
-			return fileType;
+		if (invertedContentTypes[pptxContentType]) {
+			fileType = "pptx";
+			doc.targets = concatArrays([
+				doc.targets,
+				invertedContentTypes[pptxContentType],
+			]);
 		}
-		for (let i = 0, len = defaults.length; i < len; i++) {
-			const def = defaults[i];
-			const contentType = def.getAttribute("ContentType");
-			if (
-				[docxContentType, docxmContentType, pptxContentType].indexOf(
-					contentType
-				) !== -1
-			) {
-				const extension = def.getAttribute("Extension");
-				// eslint-disable-next-line no-loop-func
-				zip.file(/./).map(({ name }) => {
-					if (
-						name.slice(name.length - extension.length - 1) === ".xml" &&
-						partNames.indexOf(name) === -1 &&
-						name !== ctXML
-					) {
-						doc.targets.push(name);
-						fileType = contentType === pptxContentType ? "pptx" : "docx";
-					}
-				});
-			}
+		if (invertedContentTypes[docxmContentType]) {
+			fileType = "docx";
+			doc.targets = concatArrays([
+				doc.targets,
+				invertedContentTypes[docxmContentType],
+			]);
 		}
 		return fileType;
 	}